split national id into its own form group
